Major refactoring of services and hooks to seperate logic and use clean architecture.
