+++ src/hooks/forms/useRecipeForm.js
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback } from "react";
 import { useNavigate } from "react-router-dom";
-import { useRecipeActions } from "./useRecipeActions";
+import { useRecipeActions } from "../data/useRecipeActions";
 import { useTranslation } from "react-i18next";
 
 export const useRecipeForm = ({ initialRecipe = null }) => {
