+++ src/main.jsx
@@ -4,7 +4,7 @@ import { createRoot } from "react-dom/client";
 
 import "../src/styles/index.css";
 import App from "./App.jsx";
-import "./i18n";
+import "./lib/i18n";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
