+++ src/pages/AuthPage/AuthPage.jsx
@@ -1,5 +1,5 @@
 import { signUp, signIn } from "../../services/auth";
-import { validateAuthForm } from "../../services/validation";
+import { validateAuthForm } from "../../utils/validation";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
