+++ src/styles/layout/containers.css
@@ -23,7 +23,7 @@
 }
 
 .card-padded {
-  padding: 1.25rem 1.25rem 0;
+  padding: 0 1.25rem 1.25rem;
   max-width: 50rem;
 }
